requirements.txt
flask
pyjwt
cryptography
flake8
pytest
coverage
requests
pytest-cov